Refactor input arch
