+++ front/src/modules/settings/profile/components/EmailField.tsx
@@ -1,13 +1,13 @@
 import { useRecoilValue } from 'recoil';
 
 import { currentUserState } from '@/auth/states/currentUserState';
-import { TextInputSettings } from '@/ui/input/text/components/TextInputSettings';
+import { TextInput } from '@/ui/input/components/TextInput';
 
 export const EmailField = () => {
   const currentUser = useRecoilValue(currentUserState);
 
   return (
-    <TextInputSettings
+    <TextInput
       value={currentUser?.email}
       disabled
       fullWidth
