+++ front/src/modules/spreadsheet-import/steps/components/ValidationStep/components/columns.tsx
@@ -5,8 +5,8 @@ import styled from '@emotion/styled';
 import { MatchColumnSelect } from '@/spreadsheet-import/components/MatchColumnSelect';
 import { Data, Fields } from '@/spreadsheet-import/types';
 import { Checkbox, CheckboxVariant } from '@/ui/input/components/Checkbox';
+import { TextInput } from '@/ui/input/components/TextInput';
 import { Toggle } from '@/ui/input/components/Toggle';
-import { TextInputSettings } from '@/ui/input/text/components/TextInputSettings';
 import { AppTooltip } from '@/ui/tooltip/AppTooltip';
 
 import { Meta } from '../types';
@@ -146,7 +146,7 @@ export const generateColumns = <T extends string>(
           }
           default:
             component = (
-              <TextInputSettings
+              <TextInput
                 value={row[columnKey] as string}
                 onChange={(value: string) => {
                   onRowChange({ ...row, [columnKey]: value });
