+++ front/src/modules/ui/field/meta-types/input/components/EmailFieldInput.tsx
@@ -1,4 +1,4 @@
-import { TextInput } from '@/ui/input/components/TextInput';
+import { TextInput } from '@/ui/field/meta-types/input/components/internal/TextInput';
 
 import { usePersistField } from '../../../hooks/usePersistField';
 import { useEmailField } from '../../hooks/useEmailField';
