+++ front/src/modules/ui/field/meta-types/input/components/internal/ProbabilityInput.tsx
