+++ front/src/modules/ui/input/components/EntityTitleDoubleTextInput.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent } from 'react';
 import styled from '@emotion/styled';
 
-import { StyledInput } from '@/ui/input/components/TextInput';
+import { StyledInput } from '@/ui/field/meta-types/input/components/internal/TextInput';
 import { ComputeNodeDimensions } from '@/ui/utilities/dimensions/components/ComputeNodeDimensions';
 
 export type EntityTitleDoubleTextInputProps = {
