+++ front/src/modules/ui/input/components/ImageInput.tsx
