+++ front/src/modules/ui/input/components/__stories__/ImageInput.stories.tsx
