+++ front/src/modules/ui/input/components/internal/date/components/__stories__/DatePicker.stories.tsx
@@ -4,11 +4,11 @@ import { userEvent, within } from '@storybook/testing-library';
 
 import { ComponentDecorator } from '~/testing/decorators/ComponentDecorator';
 
-import { DatePicker } from '../DatePicker';
+import { InternalDatePicker } from '../InternalDatePicker';
 
-const meta: Meta<typeof DatePicker> = {
-  title: 'UI/Input/DatePicker',
-  component: DatePicker,
+const meta: Meta<typeof InternalDatePicker> = {
+  title: 'UI/Input/InternalDatePicker',
+  component: InternalDatePicker,
   decorators: [ComponentDecorator],
   argTypes: {
     date: { control: 'date' },
@@ -17,7 +17,7 @@ const meta: Meta<typeof DatePicker> = {
 };
 
 export default meta;
-type Story = StoryObj<typeof DatePicker>;
+type Story = StoryObj<typeof InternalDatePicker>;
 
 export const Default: Story = {};
