+++ front/src/modules/ui/input/types/InputHotkeyScope.ts
