+++ front/src/modules/workspace/components/WorkspaceInviteLink.tsx
@@ -3,7 +3,7 @@ import styled from '@emotion/styled';
 
 import { Button } from '@/ui/button/components/Button';
 import { IconCopy, IconLink } from '@/ui/icon';
-import { TextInputSettings } from '@/ui/input/text/components/TextInputSettings';
+import { TextInput } from '@/ui/input/components/TextInput';
 import { useSnackBar } from '@/ui/snack-bar/hooks/useSnackBar';
 
 const StyledContainer = styled.div`
@@ -31,7 +31,7 @@ export const WorkspaceInviteLink = ({
   return (
     <StyledContainer>
       <StyledLinkContainer>
-        <TextInputSettings value={inviteLink} disabled fullWidth />
+        <TextInput value={inviteLink} disabled fullWidth />
       </StyledLinkContainer>
       <Button
         Icon={IconLink}
